Move news-flash API functions into separate module.
